requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-openai
langchain-community
pypdf
faiss-cpu
prophet
plotly